refactor: Move toaster related files
