+++ src/components/Organisms/Toast/Toaster/index.tsx
@@ -1,0 +1,2 @@
+export { Toaster } from './Toaster.component';
+export { reducer, toast, useToast } from './useToast';
