--- src/components/Templates/Toaster/index.tsx
@@ -1,1 +1,0 @@
-export { Toaster } from './Toaster.component'
